test_cos/requirements.txt
requests
oss2
qcloud_cos_v4